Changed the shuffle behavior in QueryGenerator a bit.
In the old implementation, it's basically randomized. Therefore, if nextRequest() is execute n times, (which n > queries.size), then it's pretty random on how many times each query in the query file is executed.

This might be a bit different of what is expected from "shuffle", which the ordering of execution might change given a queries list, however if nextRequest() is executed n times, the # of each query picked should be n / queries.size or +1 of that, instead of being random.

Shuffle gives more consistent test results yet avoid certain repeated pattern causing inaccurate/incomplete test results.

This change basically shuffles the list instead of randomized